linalg: rename svd mod and add to kornia parent
