+++ crates/kornia-linalg/src/lib.rs
@@ -2,4 +2,4 @@
 #![doc = env!("CARGO_PKG_DESCRIPTION")]
 
 /// Module to calculate SVD of a 3x3 matrix
-pub mod linalg;
+pub mod svd;
