+++ crates/kornia/Cargo.toml
@@ -17,13 +17,14 @@ gstreamer = ["kornia-io/gstreamer"]
 turbojpeg = ["kornia-io/turbojpeg"]
 
 [dependencies]
-kornia-tensor.workspace = true
-kornia-tensor-ops.workspace = true
+kornia-3d = { workspace = true }
+kornia-icp = { workspace = true }
 kornia-image.workspace = true
 kornia-imgproc.workspace = true
 kornia-io = { workspace = true, features = [] }
-kornia-3d = { workspace = true }
-kornia-icp = { workspace = true }
+kornia-linalg = { workspace = true }
+kornia-tensor.workspace = true
+kornia-tensor-ops.workspace = true
 
 [lib]
 doctest = false
